Split up functions into seperate files
